requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
tensorboard>=2.4.0
Pillow>=8.0.0 